Flytt typer for aktivitet læremidler til egen fil
